backend/foodgram/requirements.txt
Django==2.2.19
django-filter==2.4.0
pytest==6.2.4
psycopg2-binary==2.8.6
sqlparse==0.3.1
pytest-pythonpath==0.7.3
pytest-django==4.4.0
gunicorn==20.0.4
djangorestframework==3.12.4
reportlab==3.6.12
requests==2.26.0
djoser==2.0.1
python-dotenv==0.19.0
django-extensions==2.2.9